+++ index.js
@@ -7,6 +7,10 @@ const RatioBox = require('./src/RatioBox');
 const PseudoElements = require('./src/PseudoElements');
 const DevTools = require('./src/DevTools');
 const GridGap = require('./src/GridGap');
+const GridLine = require('./src/GridLine');
+const Setup = require('./src/Setup');
+const ColorTokens = require('./src/ColorTokens');
+const ApplyColorVariables = require('./src/ApplyColorVariables');
 
 module.exports = {
   Container,
@@ -17,5 +21,9 @@ module.exports = {
   RatioBox,
   PseudoElements,
   DevTools,
-  GridGap
+  GridGap,
+  GridLine,
+  Setup,
+  ColorTokens,
+  ApplyColorVariables
 };