+++ src/Container/index.js
@@ -1,61 +1,22 @@
-const _ = require('lodash');
-const getFirstBp = require('../util/getFirstBp');
-
-module.exports = function({ addComponents, theme }) {
-  const widths = theme('mainColWidths', {});
-  const outerGutters = theme('outerGutters', {});
-  const className = '.container';
-  const firstBp = getFirstBp(theme);
-  let prevGutter = -1;
-  let prevWidth = -1;
-
-  const containerStyles = _.map(widths, (width, bp) => {
-    let styles = {};
-    let gutter = outerGutters[bp];
-
-    if (prevWidth === width && prevGutter === gutter) {
-      return;
-    }
-
-    prevGutter = gutter;
-    prevWidth = width;
-
-    if (width === 'auto' || width === 'fluid') {
-      styles = {
-        width: `calc(100% - (${gutter} * 2))`
-      };
-    } else {
-      styles = {
-        width: width
-      };
-    }
-
-    if (bp === firstBp) {
-      return {
-        [className]: {
-          ...styles,
-          'margin-right': 'auto',
-          'margin-left': 'auto'
-        }
-      };
-    } else {
-      return {
-        [`@screen ${bp}`]: {
-          [className]: {
-            ...styles
-          }
-        }
-      };
+module.exports = function({ addComponents }) {
+  addComponents({
+    '.container': {
+      'width': 'calc(var(--container-width, 100%) - (2 * var(--outer-gutter, 0)))',
+      'margin-right': 'auto',
+      'margin-left': 'auto'
     }
   });
-
-  const nestedContainer = {
-    [`${className} ${className}`]: {
-      width: 'auto'
+  addComponents({
+    '.container .container': {
+      'width': 'auto'
     }
-  };
-
-  containerStyles.push(nestedContainer);
-
-  addComponents(containerStyles);
+  });
+  addComponents({
+    '.breakout': {
+      'position': 'relative',
+      'left': '50%',
+      'width': '100vw',
+      'margin-left': '-50vw'
+    }
+  });
 };