+++ src/GridGap/index.js
@@ -1,47 +1,19 @@
-require('dotenv').config();
-const getFirstBp = require('../util/getFirstBp');
-
 module.exports = function({ addBase, theme }) {
-  const innerGutters = theme('innerGutters', {});
-  const firstBp = getFirstBp(theme);
-  const gridGaps = [];
-  let prevGutter = -1;
 
-  Object.entries(innerGutters).forEach(bp => {
-    const [name, gutter] = bp;
-
-    if (gutter !== prevGutter) {
-
-      prevGutter = gutter;
-
-      let styles = {
-        '.gap-gutter': {
-          'grid-gap': gutter,
-          'gap': gutter
-        },
-        '.gap-y-gutter': {
-          'grid-row-gap': gutter,
-          'row-gap': gutter
-        },
-        '.gap-x-gutter': {
-          'grid-column-gap': gutter,
-          'column-gap': gutter
-        }
-      };
-
-      if (name === firstBp) {
-        gridGaps.push(styles);
-      } else {
-        gridGaps.push({
-          [`@screen ${name}`]: {
-            ...styles
-          }
-        });
-      }
+  let styles = {
+    '.gap-gutter': {
+      'grid-gap': 'var(--inner-gutter)',
+      'gap': 'var(--inner-gutter)'
+    },
+    '.gap-y-gutter': {
+      'grid-row-gap': 'var(--inner-gutter)',
+      'row-gap': 'var(--inner-gutter)'
+    },
+    '.gap-x-gutter': {
+      'grid-column-gap': 'var(--inner-gutter)',
+      'column-gap': 'var(--inner-gutter)'
     }
-  });
+  };
 
-  addBase(gridGaps);
+  addBase(styles);
 };
-
-